add validation and clean code

+++ src/DeleteReserveRoom.java
@@ -1,5 +1,6 @@
 import java.sql.Connection;
 import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Scanner;
 
@@ -19,21 +20,35 @@ public class DeleteReserveRoom {
             System.out.println("Enter Room Number: ");
             int roomNumber = scn.nextInt();
 
-            String sql = "DELETE FROM reservation WHERE room_number=?";
-            PreparedStatement stm = conn.prepareStatement(sql);
+            String validationQuery = "SELECT room_number FROM reservation WHERE room_number = ?";
+            PreparedStatement validationStm = conn.prepareStatement(validationQuery);
 
-            stm.setInt(1, roomNumber);
-            int result = stm.executeUpdate();
-            if(result > 0){
+            validationStm.setInt(1, roomNumber);
 
-                System.out.println("Reserve Room "+ roomNumber +" Deleted Successfully");
+            ResultSet rs = validationStm.executeQuery();
+            if(rs.next()){
+
+                String sql = "DELETE FROM reservation WHERE room_number=?";
+                PreparedStatement stm = conn.prepareStatement(sql);
+
+                stm.setInt(1, roomNumber);
+                int result = stm.executeUpdate();
+
+                if(result > 0){
+
+                    System.out.println("Reserve Room "+ roomNumber +" Deleted Successfully");
+
+                }else{
+
+                    System.out.println( roomNumber + " Room delete failed");
+
+                }
 
             }else{
-
-                System.out.println( roomNumber + " Room delete failed");
-
+                System.out.println("Invalid Room Number");
             }
 
+            rs.close();
 
         }catch(SQLException e){
